+++ examples/test_data_gen/src/construct_test_data_flow/generate_ground_truth_prompt.jinja2
