+++ examples/test_data_gen/src/test_data_gen_pipeline/config.ini
@@ -1,0 +1,17 @@
+subscription_id = "96aede12-2f73-41cb-b983-6d11a904839b"
+resource_group = "promptflow"
+; cspell: ignore yaopfeus
+workspace_name = "yaopfeus"
+aml_cluster = "cpu-cluster"
+; cspell: ignore luyao
+should_skip_doc_split = False
+documents_folder = "C:\Users\chesi\dcmcode\github-microsoft-promptflow\promptflow\examples\test_data_gen\test_docs"
+document_chunk_size = 128
+document_nodes_file_path = "C:\Users\chesi\dcmcode\github-microsoft-promptflow\promptflow\examples\test_data_gen\text-chunk-data-100.jsonl"
+; test data gen flow configs
+connection_name = "azure_open_ai_connection"
+flow_path = "C:\Users\chesi\dcmcode\github-microsoft-promptflow\promptflow\examples\test_data_gen\construct_test_data_flow\flow.dag.yaml"
+; Parallel run step configs
+prs_instance_count = 2
+prs_mini_batch_size = "10kb"
+prs_max_concurrency_per_instance = 10
