--- examples/test_data_gen/test_data_gen_pipeline/config.ini
@@ -1,17 +1,0 @@
-subscription_id = "96aede12-2f73-41cb-b983-6d11a904839b"
-resource_group = "promptflow"
-; cspell: ignore yaopfeus
-workspace_name = "yaopfeus"
-aml_cluster = "cpu-cluster"
-; cspell: ignore luyao
-should_skip_doc_split = True
-documents_folder = "D:\proj\github\ms\promptflow\docs\how-to-guides"
-document_chunk_size = 1024
-document_nodes_file_path = "D:\proj\github\ms\promptflow\examples\test_data_gen\document-nodes-test.jsonl"
-; test data gen flow configs
-connection_name = "azure_open_ai_connection"
-flow_path = "D:\proj\github\ms\promptflow\examples\test_data_gen\construct_test_data_flow\flow.dag.yaml"
-; Parallel run step configs
-prs_instance_count = 2
-prs_mini_batch_size = "10kb"
-prs_max_concurrency_per_instance = 10
